app/requirements.txt
paho-mqtt
growattServer